fix(objects): moves builtelems in correct namespace
