+++ Objects/Objects/BuiltElements/Brace.cs
@@ -4,11 +4,11 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace Objects
+namespace Objects.BuiltElements
 {
   public class Brace : Element
   {
-    
+
     public Brace()
     {
