+++ Objects/Objects/BuiltElements/Duct.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace Objects
+namespace Objects.BuiltElements
 {
   public class Duct : Element
   {
