+++ Objects/Objects/BuiltElements/Element.cs
@@ -4,13 +4,13 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace Objects
+namespace Objects.BuiltElements
 {
   public class Element : Base
   {
     public Mesh displayMesh { get; set; } = new Mesh();
 
-    public string linearUnits { get; set; } 
+    public string linearUnits { get; set; }
   }
 
   public interface IRevitElement
