+++ Objects/Objects/BuiltElements/GridLine.cs
@@ -4,7 +4,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace Objects
+namespace Objects.BuiltElements
 {
   public class GridLine : Base
   {
