+++ Objects/Objects/BuiltElements/Topography.cs
@@ -4,14 +4,14 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace Objects
+namespace Objects.BuiltElements
 {
   public class Topography : Base
   {
     public Mesh baseGeometry { get; set; } = new Mesh();
     public Topography()
     {
-      
+
     }
 
   }
